User-Pass
This program is designed to generate a customized username and password based on the user's first and last name inputs. The username is created by combining a portion of the first name with the initial of the last name, while the password is derived by rearranging the characters of the username.